Made everything super smooth

+++ src/main/java/ellpeck/actuallyadditions/blocks/render/RenderSmileyCloud.java
@@ -46,18 +46,18 @@ public class RenderSmileyCloud extends RenderTileEntity{
                 theCloud.flyHeight = tile.getWorldObj().rand.nextInt(30)+30;
             }
             int bobHeight = theCloud.flyHeight;
-            long theTime = Minecraft.getSystemTime();
-            long time = theTime/50;
+            double theTime = Minecraft.getSystemTime();
+            double time = theTime/50;
 
             if(time-bobHeight >= theCloud.lastFlyHeight){
                 theCloud.lastFlyHeight = time;
             }
 
             if(time-(bobHeight/2) >= theCloud.lastFlyHeight){
-                GL11.glTranslated(0, ((double)time-theCloud.lastFlyHeight)/300, 0);
+                GL11.glTranslated(0, (time-theCloud.lastFlyHeight)/300, 0);
             }
             else{
-                GL11.glTranslated(0, -((double)time-theCloud.lastFlyHeight)/300+(double)bobHeight/300, 0);
+                GL11.glTranslated(0, -(time-theCloud.lastFlyHeight)/300+(double)bobHeight/300, 0);
             }
 
             GL11.glTranslatef((float)x+0.5F, (float)y-0.5F, (float)z+0.5F);
